additiona unit test to cover show vlan
